+++ NEWS.md
@@ -5,6 +5,7 @@
 * New function `tree_hulls` that computes a convex or concave hull for each segmented tree.
 * New option `stop_early` to process a catalog until the end anyway or stop before the end if an error is raised.
 * New function `catalog_retile` supersede the function `catalog_reshape` being able to do the same and much more.
+* New function `lastrees_ptree` with Vega et al. (2014) algorithm.
 
 #### ENHANCEMENTS